requirements.txt
anyio==3.7.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.2
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.21.0
google-generativeai==0.1.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
idna==3.4
mangum==0.17.0
markdown-it-py==2.2.0
mdurl==0.1.2
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
Pygments==2.15.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
rich==13.3.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
sseclient==0.0.27
starlette==0.27.0
typing_extensions==4.6.2
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
